Fixed weird naming convention
