+++ frontend-Idata2301/code/app/src/components/cards/searched-flights/FlightsContainer.jsx
@@ -1,53 +1,18 @@
-import React, { useState, useEffect } from "react";
-import FlightCard from "./FlightCard";
-import "./flightsContainer.css"; // Import the container styles
+import React from "react";
+import NewFlightCard from "./FlightCard";
+import "./flightsContainer.css"; 
 
-const FlightsContainer = () => {
-  const [flights, setFlights] = useState([]);
-
-  useEffect(() => {
-    const fetchFlights = async () => {
-      const fakeData = [
-        {
-          airline: "Delta Airlines",
-          flightNumber: "DL123",
-          departure: "New York (JFK)",
-          arrival: "Los Angeles (LAX)",
-          price: 320,
-        },
-        {
-          airline: "American Airlines",
-          flightNumber: "AA456",
-          departure: "Chicago (ORD)",
-          arrival: "Miami (MIA)",
-          price: 280,
-        },
-        {
-            airline: "American Airlines",
-            flightNumber: "AA456",
-            departure: "Chicago (ORD)",
-            arrival: "Miami (MIA)",
-            price: 280,
-        },
-        {
-            airline: "American Airlines",
-            flightNumber: "AA456",
-            departure: "Chicago (ORD)",
-            arrival: "Miami (MIA)",
-            price: 280,
-        },
-      ];
-      setFlights(fakeData);
-    };
-
-    fetchFlights();
-  }, []);
+const FlightsContainer = ({ flights }) => {  // Receive flights as a prop
 
   return (
     <div className="flights-container flights-grid">
-      {flights.map((flight, index) => (
-        <FlightCard key={index} flight={flight} />
-      ))}
+      {flights.length > 0 ? (
+        flights
+          .filter(flight => !flight.isHidden)  // Filter out flights where isHidden is true
+          .map((flight, index) => <NewFlightCard key={index} flight={flight} />)
+      ) : (
+        <p>No flights found fitting given criteria.</p>
+      )}
     </div>
   );
 };
