--- frontend-Idata2301/code/app/src/components/cards/searched-flights/newFlightCard.css
@@ -1,66 +1,0 @@
-.new-flight-card {
-    background: #ffffff;
-    border-radius: 12px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    padding: 20px;
-    width: 300px;
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    text-align: center;
-  }
-  
-  .new-flight-card h2 {
-    font-size: 18px;
-    color: #2c3e50;
-    margin-bottom: 5px;
-  }
-  
-  .new-class-name {
-    font-size: 14px;
-    font-weight: bold;
-    color: #16a085;
-  }
-  
-  .new-flight-details {
-    display: flex;
-    justify-content: space-between;
-    padding: 10px 0;
-  }
-  
-  .label {
-    font-size: 12px;
-    color: #7f8c8d;
-  }
-  
-  .new-date, .new-seats {
-    font-size: 14px;
-    font-weight: bold;
-    color: #34495e;
-  }
-  
-  .new-price-section {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 10px;
-  }
-  
-  .new-price {
-    font-size: 16px;
-    font-weight: bold;
-    color: #e74c3c;
-  }
-  
-  .new-discount {
-    font-size: 14px;
-    color: #27ae60;
-    font-weight: bold;
-  }
-  
-  .new-provider {
-    font-size: 12px;
-    color: #7f8c8d;
-    margin-top: 5px;
-  }
-
